Redux: Reducers and Actions
My first step before plunging into using Redux is to define and
setup a set of Reducers and Actions.

Reducers are essentially a set of pure functions that respond to
actions that dispatched onto out store.

Actions are functions that let us iterate to our next version of
our store's state.