Fixing "Type" regex, and cleaning a critical bug.
Initially I was instantiating a bunch of regex objects, so that
I could reuse them and be a "tiny" bit more performant. However,
by doing that I opened up an unexpected consequence of giving
inconsistent results each time `myRegex.exec` was ran.

So, this PR moves the regex initialization into the module's `export`,
and I've cleaned up the regex a little so that it takes into account
monsters with multiple types.